Conditionally set he image data format to channels_last
On my configuration with versions:

```
TF version: 2.10.0
Keras version: 2.10.0
```

I had to make this simple change in order to have this notebook work. Otherwise, creating the VGG19 model would fail indicating that the channel was incorrectly positioned.

I would even argue that perhaps this configuration should not be commented, but I'll leave it to you to decide. Let me know if I should change.